requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
httpx
pydantic
requests